M32040: Missing word or tag
Hello, @MichaelNorman,  
Translator has reported possible source content issue. 
Description:´The tag is missing in the source (whose ? to set). For the time being we inserted "BoxView" because we did not want to deliver an incomplete sentence. Please advise
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.