M32047: Possible wrong tag in source
Hello, @MichaelNorman, @lobrien,
Translator has reported possible source content issue. 
Description: Could it be possibe that "this" has been tagged wrongly and that it is just a normal word and not an element/classe/attribute?
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.